@@ -1,6 +1,6 @@
 {
   "name": "@studiometa/js-toolkit-demo",
-  "version": "2.12.0",
+  "version": "3.0.0-alpha.0",
   "private": true,
   "type": "commonjs",
   "scripts": {
@@ -11,15 +11,15 @@
   "dependencies": {
     "@studiometa/eslint-config": "^3.1.3",
     "@studiometa/stylelint-config": "^3.0.2",
-    "@studiometa/tailwind-config": "^2.0.2",
-    "@studiometa/ui": "^0.2.40",
+    "@studiometa/tailwind-config": "^2.1.0",
+    "@studiometa/ui": "^0.2.47",
     "@studiometa/webpack-config": "^5.3.0",
     "@studiometa/webpack-config-preset-prototyping": "^5.3.0",
-    "eslint": "^8.54.0",
-    "postcss": "^8.4.31",
+    "eslint": "^8.57.0",
+    "postcss": "^8.4.38",
     "prettier": "^2.8.8",
     "stylefire": "^7.0.3",
     "stylelint": "^15.11.0",
-    "tailwindcss": "^3.3.5"
+    "tailwindcss": "^3.4.3"
   }
 }
