@@ -1,69 +1,24 @@
-const path = require('path');
-const glob = require('fast-glob');
-const esbuild = require('esbuild');
+import { resolve, dirname } from 'node:path';
+import glob from 'fast-glob';
+import esbuild from 'esbuild';
 
-const entryPoints = glob.sync(
-  ['packages/js-toolkit/**/*.js', 'packages/js-toolkit/**/*.ts', '!**/node_modules/**'],
-  {
-    cwd: path.resolve(__dirname, '..'),
-  },
-);
+const root = resolve(dirname(new URL(import.meta.url).pathname), '..');
 
-const outdir = path.resolve(__dirname, '../dist');
-
-const defaultOptions = {
-  entryPoints,
+const options = {
+  entryPoints: glob.globSync(['packages/js-toolkit/**/*.ts', '!**/node_modules/**'], {
+    cwd: root,
+  }),
   write: true,
-  outdir,
+  outdir: resolve(root, 'dist'),
   target: 'esnext',
+  format: 'esm',
   sourcemap: true,
 };
 
-/**
- * Build with esbuild.
- *
- * @param   {require('esbuild').BuildOptions} opts
- * @returns {Promise<void>}
- */
-async function build(opts) {
-  console.log(`Building ${opts.format}...`);
-  const { errors, warnings } = await esbuild.build({
-    ...defaultOptions,
-    ...opts,
-  });
+console.log(`Building...`);
+const { errors, warnings } = await esbuild.build(options);
 
-  errors.forEach(console.error);
-  warnings.forEach(console.warn);
-  console.log(`Done building ${opts.format}!`);
-}
+errors.forEach(console.error);
+warnings.forEach(console.warn);
 
-build({
-  format: 'esm',
-});
-
-build({
-  format: 'cjs',
-  bundle: true,
-  outExtension: { '.js': '.cjs' },
-  footer: {
-    // Fix `export default {}` assigned to `module.exports.default`
-    js: 'if (module.exports.default) module.exports = module.exports.default;',
-  },
-  plugins: [
-    // @see https://github.com/evanw/esbuild/issues/622#issuecomment-769462611
-    {
-      name: 'change-extension-to-cjs',
-      setup(builder) {
-        // eslint-disable-next-line consistent-return
-        builder.onResolve({ filter: /.*/ }, (args) => {
-          if (args.importer) {
-            return {
-              path: args.path.replace(/\.js$/, '.cjs'),
-              external: true,
-            };
-          }
-        });
-      },
-    },
-  ],
-});
+console.log(`Done building!`);
