@@ -12,6 +12,9 @@ export default class AnimateTestMultiple extends AnimateTest {
     ...AnimateTest.config,
     name: 'AnimateTestMultiple',
     refs: ['targets[]', 'start', 'pause', 'play', 'finish', 'progress'],
+    options: {
+      stagger: Number,
+    },
   };
 
   mounted() {
@@ -24,34 +27,20 @@ export default class AnimateTestMultiple extends AnimateTest {
     });
 
     const options = {
+      smooth: this.$options.smooth,
+      stagger: this.$options.stagger,
       duration: this.$options.duration,
       easing: ease[this.$options.easing],
     };
 
-    const animations = this.$refs.targets.map((target, index) => {
-      return animate(target, steps, {
-        ...options,
-        onProgress:
-          index === 0
-            ? (progress) => {
-                this.$refs.progress.value = progress;
-              }
-            : undefined,
-      });
+    console.log(options);
+
+    this.animate = animate(this.$refs.targets, steps, {
+      ...options,
+      onProgress: (progress) => {
+        this.$refs.progress.value = progress;
+      },
     });
-
-    const delegate =
-      (action) =>
-      (...args) => {
-        animations.forEach((animation) => animation[action](...args));
-      };
-
-    this.animate = {
-      start: delegate('start'),
-      pause: delegate('pause'),
-      play: delegate('play'),
-      finish: delegate('finish'),
-      progress: delegate('progress'),
-    };
+    this.animate.progress(this.$refs.progress.valueAsNumber);
   }
 }
