@@ -1,3 +1,26 @@
-export type Features = Record<'asyncChildren', boolean>;
+export type Features = {
+  asyncChildren: boolean;
+  breakpoints: Record<string, string>;
+};
 
-export const features = new Map<keyof Features, boolean>([['asyncChildren', false]]);
+interface FeaturesMap extends Map<keyof Features, Features[keyof Features]> {
+  get<T extends keyof Features>(key: T): Features[T];
+  set<T extends keyof Features>(key: T, value: Features[T]): this;
+}
+
+export const features = new Map<keyof Features, Features[keyof Features]>([
+  ['asyncChildren', false],
+  [
+    'breakpoints',
+    {
+      xxs: '0rem',
+      xs: '30rem', // 480px
+      s: '48rem', // 768px
+      m: '64rem', // 1024px
+      l: '80rem', // 1280px
+      xl: '90rem', // 1440px
+      xxl: '120rem', // 1920px
+      xxxl: '160rem', // 2560px
+    },
+  ],
+]) as FeaturesMap;
