@@ -41,7 +41,7 @@ export function withIntersectionObserver<S extends Base>(
       ...BaseClass.config,
       name: `${BaseClass.config.name}WithIntersectionObserver`,
       options: {
-        ...(BaseClass.config?.options || {}),
+        ...BaseClass.config?.options,
         intersectionObserver: Object,
       },
       emits: ['intersected'],
