@@ -1,6 +1,6 @@
-import { lerp, map } from '../math/index.js';
+import { clamp01, lerp, map, mean } from '../math/index.js';
 import { isDefined, isFunction, isNumber } from '../is.js';
-import transform, { TRANSFORM_PROPS } from './transform.js';
+import { transform, TRANSFORM_PROPS } from './transform.js';
 import { domScheduler as scheduler } from '../scheduler.js';
 import { tween, normalizeEase } from '../tween.js';
 // eslint-disable-next-line import/extensions
@@ -8,17 +8,23 @@ import { eachElements } from './utils.js';
 import { startsWith } from '../string/index.js';
 import type { TransformProps } from './transform.js';
 import type { EasingFunction } from '../math/index.js';
-import type { BezierCurve, TweenOptions } from '../tween.js';
+import type { BezierCurve, Tween, TweenOptions } from '../tween.js';
 
 export type CSSCustomPropertyName = `--${string}`;
 
-export type Keyframe = TransformProps & {
-  opacity?: number;
-  transformOrigin?: string;
-  easing?: EasingFunction | BezierCurve;
-  offset?: number;
-  [key: CSSCustomPropertyName]: number;
+type KeyframeTransforms<Type> = {
+  [Property in keyof Type]: number | [number, string];
 };
+
+export type Keyframe = Partial<
+  KeyframeTransforms<TransformProps> & {
+    opacity: number;
+    transformOrigin: string;
+    easing: EasingFunction | BezierCurve;
+    offset: number;
+    [key: CSSCustomPropertyName]: number;
+  }
+>;
 
 export type NormalizedKeyframe = Keyframe & {
   easing: EasingFunction;
@@ -26,13 +32,28 @@ export type NormalizedKeyframe = Keyframe & {
   vars: string[];
 };
 
-export type Animate = {
-  start: () => void;
-  pause: () => void;
-  play: () => void;
-  finish: () => void;
-  progress: (value?: number) => number;
-};
+export interface AnimateOptions extends Omit<TweenOptions, 'duration'> {
+  /**
+   * The duration of the tween, in seconds.
+   * When using multiple targets, it can be a function which will be
+   * called with the current target and its index and should return
+   * the duration in seconds for this element.
+   *
+   * Defaults to `1`.
+   */
+  duration?: number | ((target: HTMLElement, index: number) => number);
+  /**
+   * Delay between the start of each target animation, in seconds.
+   * When using multiple targets, it can be a function which will be called
+   * with the current target and its index and should return the delay
+   * in seconds for this element's tween.
+   *
+   * Defaults to `0`.
+   */
+  stagger?: number | ((target: HTMLElement, index: number) => number);
+}
+
+export type Animate = Tween;
 
 let id = 0;
 const running = new WeakMap();
@@ -123,22 +144,23 @@ function render(
 
     let customProperties: false | [string, number][] = false;
     if (isDefined(from.vars) && isDefined(to.vars)) {
-      customProperties = from.vars.map((customPropertyName) => {
-        const customPropertyValue = lerp(
-          from[customPropertyName],
-          to[customPropertyName],
-          stepProgress,
-        );
-        return [customPropertyName, customPropertyValue];
-      });
+      customProperties = [];
+      for (const customPropertyName of from.vars) {
+        customProperties.push([
+          customPropertyName,
+          lerp(from[customPropertyName], to[customPropertyName], stepProgress),
+        ]);
+      }
     }
 
-    const props = Object.fromEntries(
-      TRANSFORM_PROPS.filter((name) => isDefined(from[name]) || isDefined(to[name])).map((name) => [
-        name,
-        transformRenderStrategies[name](element, from[name], to[name], stepProgress),
-      ]),
-    );
+    const props = {};
+
+    for (const name of TRANSFORM_PROPS) {
+      if (isDefined(from[name]) || isDefined(to[name])) {
+        props[name] = transformRenderStrategies[name](element, from[name], to[name], stepProgress);
+      }
+    }
+
     scheduler.write(() => {
       if (opacity !== false) {
         // @ts-ignore
@@ -200,9 +222,9 @@ function singleAnimate(
 
   const controls = tween(callback, {
     ...options,
-    onStart() {
+    onStart(progress) {
       if (isFunction(options.onStart)) {
-        options.onStart();
+        options.onStart(progress);
       }
       // Stop running instances
       const runningKeys = running.get(element);
@@ -218,28 +240,88 @@ function singleAnimate(
   return controls;
 }
 
+type Duration = number;
+type Delay = number;
+type DurationWithDelay = number;
+
 /**
  * Animate one or more elements.
  */
 export function animate(
   elementOrElements: HTMLElement | HTMLElement[] | NodeListOf<HTMLElement>,
   keyframes: Keyframe[],
-  options: TweenOptions = {},
+  options: AnimateOptions = {},
 ): Animate {
-  const controls = eachElements(elementOrElements, (element) =>
-    singleAnimate(element, keyframes, options),
-  );
+  if (elementOrElements instanceof HTMLElement) {
+    return singleAnimate(elementOrElements, keyframes, {
+      ...options,
+      duration: isFunction(options.duration)
+        ? options.duration(elementOrElements, 0)
+        : options.duration,
+    });
+  }
 
-  const delegate =
-    (key) =>
+  const stagger = options.stagger ?? 0;
+  const staggerIsFunction = isFunction(stagger);
+  const durationFn = isFunction(options.duration) ? options.duration : null;
+  const progressFn = isFunction(options.onProgress) ? options.onProgress : null;
+  const progresses: number[] = [];
+  const timings: [Duration, Delay, DurationWithDelay][] = [];
+  let duration = 0;
+  let previousTimings = [0, 0, 0];
+
+  const controls = eachElements(elementOrElements, (element, index) => {
+    const delay = staggerIsFunction ? stagger(element, index) : stagger * index;
+    const itemOptions = { ...options, delay } as TweenOptions;
+
+    if (durationFn) {
+      itemOptions.duration = durationFn(element, index);
+    } else if (!isDefined(itemOptions.duration)) {
+      itemOptions.duration = 1;
+    }
+
+    timings[index] = [itemOptions.duration, delay, itemOptions.duration + delay];
+
+    if (timings[index][2] > previousTimings[2]) {
+      // eslint-disable-next-line prefer-destructuring
+      duration = timings[index][2];
+    }
+
+    previousTimings = timings[index];
+
+    progresses[index] = 0;
+    itemOptions.onProgress = (itemProgress) => {
+      progresses[index] = itemProgress;
+      if (progressFn) {
+        progressFn(mean(progresses));
+      }
+    };
+
+    return singleAnimate(element, keyframes, itemOptions);
+  });
+
+  const delegate = (key) =>
     // eslint-disable-next-line consistent-return
-    (...args) => {
-      if (key === 'progress' && args.length === 0) {
-        return controls[0][key]();
+    function delegated() {
+      if (key === 'progress') {
+        if (arguments.length === 1) {
+          // eslint-disable-next-line prefer-rest-params
+          const newProgress = arguments[0];
+          const newTime = lerp(0, duration, newProgress);
+          for (const [index, control] of controls.entries()) {
+            const controlProgress = clamp01(
+              map(newTime, timings[index][1], timings[index][2], 0, 1),
+            );
+            control.progress(controlProgress);
+          }
+        }
+
+        return mean(progresses);
       }
 
       for (const control of controls) {
-        control[key](...args);
+        // eslint-disable-next-line prefer-rest-params
+        control[key].call(null, arguments);
       }
     };
 
