@@ -3,5 +3,6 @@ export { add as addClass, remove as removeClass, toggle as toggleClass } from '.
 export { add as addStyle, remove as removeStyle } from './styles.js';
 export { default as getOffsetSizes } from './getOffsetSizes.js';
 export { default as matrix } from './matrix.js';
-export { default as transform } from './transform.js';
+export { transform } from './transform.js';
 export { default as transition } from './transition.js';
+export type { TransformProps } from './transform.js';
