@@ -1,9 +1,9 @@
 export function eachElements<S extends Node = Node, T = void>(
   elementOrElements: S | S[] | NodeListOf<S>,
-  callback: (element: S, index?: number, elements?: S[]) => T,
+  callback: (element: S, index: number, elements: S[]) => T,
 ): T[] {
   if (elementOrElements instanceof Node) {
-    return [callback(elementOrElements)];
+    return [callback(elementOrElements, 0, [elementOrElements])];
   }
 
   return Array.from(elementOrElements).map(callback);
