@@ -1,16 +1,15 @@
 export { default as debounce } from './debounce.js';
-export { default as focusTrap } from './focusTrap.js';
+export * from './trapFocus.js';
 export { default as keyCodes } from './keyCodes.js';
 export { default as memoize } from './memoize.js';
 export { nextFrame } from './nextFrame.js';
-export { default as nextTick } from './nextTick.js';
-export { default as nextMicrotask } from './nextMicrotask.js';
+export { nextTick } from './nextTick.js';
+export { nextMicrotask } from './nextMicrotask.js';
 export { default as throttle } from './throttle.js';
-export { default as scrollTo } from './scrollTo.js';
+export { scrollTo } from './scrollTo.js';
 export { default as getComponentResolver } from './getComponentResolver.js';
 export * from './is.js';
 export * from './has.js';
-// eslint-disable-next-line import/extensions
 export * from './css/index.js';
 export {
   objectToURLSearchParams,
@@ -27,3 +26,4 @@ export { Queue } from './Queue.js';
 export { SmartQueue } from './SmartQueue.js';
 export * from './dom/index.js';
 export * from './wait.js';
+export * from './random.js';
