@@ -1,3 +1,6 @@
+export * as ease from './ease.js';
+export * from './createEases.js';
+export * from './ease.js';
 export { default as clamp } from './clamp.js';
 export { default as clamp01 } from './clamp01.js';
 export { default as damp } from './damp.js';
@@ -5,6 +8,4 @@ export { default as inertiaFinalValue } from './inertiaFinalValue.js';
 export { default as lerp } from './lerp.js';
 export { default as map } from './map.js';
 export { default as round } from './round.js';
-export * from './ease.js';
-export * as ease from './ease.js';
-export * from './createEases.js';
+export { mean } from './mean.js';
