@@ -1,11 +1,13 @@
 import { isFunction } from './is.js';
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type Callback = (...args: any[]) => any;
+
 /**
  * Wait for the next microtask.
  */
-export default async function nextMicrotask<T extends () => unknown>(
-  fn?: T,
-): Promise<T extends () => unknown ? ReturnType<T> : void> {
-  // @ts-ignore
-  return Promise.resolve().then(() => isFunction(fn) && fn());
+export async function nextMicrotask(): Promise<void>;
+export async function nextMicrotask<T extends Callback>(callback?: T): Promise<ReturnType<T>>;
+export async function nextMicrotask<T extends Callback>(callback?: T): Promise<ReturnType<T>> {
+  return Promise.resolve().then(() => isFunction(callback) && callback());
 }
